+++ components/ui/NavBar.tsx
@@ -1,5 +1,5 @@
 import { router, usePathname, useSegments } from "expo-router";
-import React, { useEffect, useRef, useState } from "react";
+import React, { useRef, useState } from "react";
 import {
   Keyboard,
   Text,
@@ -7,35 +7,21 @@ import {
   TouchableOpacity,
   View,
 } from "react-native";
-import { IconSymbol } from "@/components/ui/IconSymbol";
-import { useColorScheme } from "@/hooks/useColorScheme";
-import { supabase } from "@/lib/supabase";
 
-// Define the structure for menu options
-interface MenuOption {
-  label: string;
-  onPress: () => void;
-}
+import { useColorScheme } from "../../hooks/useColorScheme";
+import { IconSymbol } from "./IconSymbol";
 
-// Default menu options for pages not explicitly configured
-const defaultMenuOptions: MenuOption[] = [
-  {
-    label: "SETTINGS",
-    onPress: () => {
-      console.log("Default settings pressed");
-      // Add your custom logic here
-    },
-  },
-  {
-    label: "HELP",
-    onPress: () => {
-      console.log("Default help pressed");
-      // Add your custom logic here
-    },
-  },
-];
-
-export function NavBar() {
+export function NavBar({
+  customTitle,
+  showBackButton,
+  onBack,
+  invertTitle,
+}: {
+  customTitle?: string;
+  showBackButton?: boolean;
+  onBack?: () => void;
+  invertTitle?: boolean;
+} = {}) {
   const colorScheme = useColorScheme();
   const pathname = usePathname();
   const segments = useSegments();
@@ -70,18 +56,27 @@ export function NavBar() {
   const [menuOpen, setMenuOpen] = useState(false);
   const inputRef = useRef<TextInput>(null);
 
-  // Close menu when pathname changes
-  useEffect(() => {
-    setMenuOpen(false);
-  }, [pathname]);
-
   const handleMenuPress = () => {
     setMenuOpen(!menuOpen);
   };
 
+  const handleOption1 = () => {
+    console.log("Option 1 pressed");
+    setMenuOpen(false);
+  };
+
+  const handleOption2 = () => {
+    console.log("Option 2 pressed");
+    setMenuOpen(false);
+  };
+
+  const handleOption3 = () => {
+    console.log("Option 3 pressed");
+    setMenuOpen(false);
+  };
+
   const handleSearchPress = () => {
     setSearchActive(true);
-    setMenuOpen(false);
     setTimeout(() => {
       inputRef.current?.focus();
     }, 100);
@@ -104,173 +99,6 @@ export function NavBar() {
       Keyboard.dismiss();
     }
   };
-
-  // Logout handler
-  const handleLogout = async () => {
-    console.log("Logout pressed");
-    try {
-      const { error } = await supabase.auth.signOut();
-      if (error) {
-        console.error("Error signing out:", error.message);
-      } else {
-        console.log("Successfully signed out");
-      }
-    } catch (error) {
-      console.error("Error during sign out:", error);
-    }
-  };
-
-  // Get menu options for current page with dynamic handlers
-  const getCurrentMenuOptions = (): MenuOption[] => {
-    const baseConfigs = {
-      "fashion:week": [
-        {
-          label: "ADD FRIENDS",
-          onPress: () => {
-            console.log("Add friends pressed on homepage");
-            // Add your custom logic here
-          },
-        },
-        {
-          label: "SETTINGS",
-          onPress: () => {
-            console.log("Settings pressed on homepage");
-            // Add your custom logic here
-          },
-        },
-        {
-          label: "ABOUT",
-          onPress: () => {
-            console.log("About pressed on homepage");
-            // Add your custom logic here
-          },
-        },
-      ],
-      COLLECTIONS: [
-        {
-          label: "CREATE COLLECTION",
-          onPress: () => {
-            console.log("Create collection pressed");
-            // Add your custom logic here
-          },
-        },
-        {
-          label: "SHARE COLLECTION",
-          onPress: () => {
-            console.log("Share collection pressed");
-            // Add your custom logic here
-          },
-        },
-        {
-          label: "SETTINGS",
-          onPress: () => {
-            console.log("Settings pressed on collections");
-            // Add your custom logic here
-          },
-        },
-      ],
-      "DROP TRACKER": [
-        {
-          label: "ADD NEW DROP",
-          onPress: () => {
-            console.log("Add new drop pressed");
-            // Add your custom logic here
-          },
-        },
-        {
-          label: "DROP HISTORY",
-          onPress: () => {
-            console.log("Drop history pressed");
-            // Add your custom logic here
-          },
-        },
-        {
-          label: "NOTIFICATIONS",
-          onPress: () => {
-            console.log("Notifications pressed");
-            // Add your custom logic here
-          },
-        },
-      ],
-      ACCOUNT: [
-        {
-          label: "EDIT PROFILE",
-          onPress: () => {
-            console.log("Edit profile pressed");
-            router.push("/(tabs)/(user)/edit-profile");
-            setMenuOpen(false);
-          },
-        },
-        {
-          label: "PRIVACY SETTINGS",
-          onPress: () => {
-            console.log("Privacy settings pressed");
-            router.push("/(tabs)/(user)/settings");
-            setMenuOpen(false);
-          },
-        },
-        {
-          label: "LOGOUT",
-          onPress: handleLogout,
-        },
-      ],
-      ARCHIVE: [
-        {
-          label: "CLEAR ARCHIVE",
-          onPress: () => {
-            console.log("Clear archive pressed");
-            // Add your custom logic here
-          },
-        },
-        {
-          label: "EXPORT DATA",
-          onPress: () => {
-            console.log("Export data pressed");
-            // Add your custom logic here
-          },
-        },
-      ],
-      "STYLE QUIZ": [
-        {
-          label: "RESTART QUIZ",
-          onPress: () => {
-            console.log("Restart quiz pressed");
-            // Add your custom logic here
-          },
-        },
-        {
-          label: "VIEW RESULTS",
-          onPress: () => {
-            console.log("View results pressed");
-            // Add your custom logic here
-          },
-        },
-      ],
-      "SEARCH RESULTS": [
-        {
-          label: "CLEAR SEARCH",
-          onPress: () => {
-            console.log("Clear search pressed");
-            // Add your custom logic here
-          },
-        },
-        {
-          label: "SAVE SEARCH",
-          onPress: () => {
-            console.log("Save search pressed");
-            // Add your custom logic here
-          },
-        },
-      ],
-    };
-
-    return (
-      baseConfigs[pageDisplayName as keyof typeof baseConfigs] ||
-      defaultMenuOptions
-    );
-  };
-
-  const currentMenuOptions = getCurrentMenuOptions();
 
   return (
     <View>
@@ -341,20 +169,16 @@ export function NavBar() {
 
       {/* Dropdown Menu */}
       {menuOpen && (
-        <View
-          className={`mx-4 mb-4 bg-transparent flex-col justify-start ${
-            pageDisplayName === "fashion:week"
-              ? "absolute top-0 left-0 right-0 z-40 mt-32"
-              : ""
-          }`}
-        >
-          {currentMenuOptions.map((option, index) => (
-            <TouchableOpacity key={index} onPress={option.onPress}>
-              <Text className="px-4 py-1 text-sm font-bold">
-                {option.label}
-              </Text>
-            </TouchableOpacity>
-          ))}
+        <View className="mx-4 mb-4 bg-transparent flex-row justify-start flex-wrap">
+          <TouchableOpacity onPress={handleOption1}>
+            <Text className="px-4 py-0 text-sm font-bold">ADD FRIENDS</Text>
+          </TouchableOpacity>
+          <TouchableOpacity onPress={handleOption2}>
+            <Text className="px-4 py-0 text-sm font-bold">SETTINGS</Text>
+          </TouchableOpacity>
+          <TouchableOpacity onPress={handleOption3}>
+            <Text className="px-4 py-0 text-sm font-bold">BLAH BLAH BLAH</Text>
+          </TouchableOpacity>
         </View>
       )}
     </View>
